universal async access
